Add commands to Toggle Sidebar or Toolbar
- Activate command palette item on mouse enter
- Use a lighter color for command palette item hover
- Adjust command palette input color in dark mode
