+++ src/notes/filters/index.ts
@@ -1,0 +1,3 @@
+import { NotesObject } from "shared/storage/schema";
+
+export const getFirstAvailableNoteName = (notes: NotesObject): string => Object.keys(notes).sort().shift() || "";
