+++ src/notes/location.ts
@@ -1,0 +1,2 @@
+export const getFocusOverride = (): boolean => new URL(window.location.href).searchParams.get("focus") === "";
+export const getActiveFromUrl = (): string => new URL(window.location.href).searchParams.get("note") || ""; // Bookmark
